LayoutTests/storage/indexeddb/legacy-constants-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: Numeric transaction modes are deprecated in IDBDatabase.transaction. Use "readonly" or "readwrite".
CONSOLE MESSAGE: Numeric transaction modes are deprecated in IDBDatabase.transaction. Use "readonly" or "readwrite".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric transaction modes are deprecated in IDBDatabase.transaction. Use "readonly" or "readwrite".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric transaction modes are deprecated in IDBDatabase.transaction. Use "readonly" or "readwrite".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric transaction modes are deprecated in IDBDatabase.transaction. Use "readonly" or "readwrite".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
CONSOLE MESSAGE: Numeric direction values are deprecated in openCursor and openKeyCursor. Use "next", "nextunique", "prev", or "prevunique".
Test that legacy direction/mode constants work

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;


openreq = indexedDB.open('legacy-constants')
db = openreq.result
verreq = db.setVersion('1')
Deleted all object stores.
store = db.createObjectStore('store')
store.createIndex('index', 'value')

populating store...
trans = db.transaction('store', IDBTransaction.READ_WRITE)
PASS trans.mode is "readwrite"
store = trans.objectStore('store');
store.put({value: 111}, 1)
store.put({value: 222}, 2)
trans = db.transaction('store', IDBTransaction.READ_ONLY)
PASS trans.mode is "readonly"
request = store.openCursor(null, IDBCursor.NEXT)
PASS cursor.direction is "next"
cursor.continue();
PASS cursor.direction is "next"
cursor.continue();
request = store.openCursor(1, IDBCursor.NEXT)
PASS cursor.direction is "next"
cursor.continue();
trans = db.transaction('store', IDBTransaction.READ_ONLY)
PASS trans.mode is "readonly"
request = store.openCursor(null, IDBCursor.NEXT_NO_DUPLICATE)
PASS cursor.direction is "nextunique"
cursor.continue();
PASS cursor.direction is "nextunique"
cursor.continue();
request = store.openCursor(1, IDBCursor.NEXT_NO_DUPLICATE)
PASS cursor.direction is "nextunique"
cursor.continue();
trans = db.transaction('store', IDBTransaction.READ_ONLY)
PASS trans.mode is "readonly"
request = store.openCursor(null, IDBCursor.PREV)
PASS cursor.direction is "prev"
cursor.continue();
PASS cursor.direction is "prev"
cursor.continue();
request = store.openCursor(1, IDBCursor.PREV)
PASS cursor.direction is "prev"
cursor.continue();
trans = db.transaction('store', IDBTransaction.READ_ONLY)
PASS trans.mode is "readonly"
request = store.openCursor(null, IDBCursor.PREV_NO_DUPLICATE)
PASS cursor.direction is "prevunique"
cursor.continue();
PASS cursor.direction is "prevunique"
cursor.continue();
request = store.openCursor(1, IDBCursor.PREV_NO_DUPLICATE)
PASS cursor.direction is "prevunique"
cursor.continue();

Verify that constants from previous version of the spec (beyond a grace period) have been removed:
PASS IDBKeyRange.SINGLE is undefined
PASS IDBKeyRange.LEFT_OPEN is undefined
PASS IDBKeyRange.RIGHT_OPEN is undefined
PASS IDBKeyRange.LEFT_BOUND is undefined
PASS IDBKeyRange.RIGHT_BOUND is undefined
PASS IDBTransaction.LOADING is undefined
PASS IDBRequest.LOADING is undefined
PASS IDBRequest.DONE is undefined
PASS successfullyParsed is true

TEST COMPLETE
